// angular/src/app/enums/one-pass-url.enum.ts
export enum OnePassUrlEnum {
  loginUser = 'LoginUser',
  getOnePassDataByUserId = 'GetOnePassDataByUserId',
  deletePassword = 'DeletePassword',
  deleteServer = 'DeleteServer',
  deleteCreditCard = 'DeleteCreditCard',
  deleteWifiPassword = 'DeleteWifiPassword',
  addOrUpdatePassword = 'AddOrUpdatePassword',
  addPasswords = 'AddPasswords',
  addOrUpdateServer = 'AddOrUpdateServer',
  addOrUpdateWifiPassword = 'AddOrUpdateWifiPassword',
  addOrUpdateCreditCard = 'AddOrUpdateCreditCard',
  sendCode = 'SendCode',
  verify = 'Verify',
  deleteDevice = 'DeleteDevice',
  register = 'Register',
  token = 'token'
}